MDL-45075 mod_forum: Move user subscription to an observer
The forum subscription function is called in some cases on forum creation
(specifically the case of an automatically subscribed forum).

Since the forum subscription requires both the course_module instance, and
the context these must both be present when subscribing users to the forum,
however these are not fully set up until after the forum has been created.

By moving the subscription to a subscriber, we are able to hook into the
module creation event after the course_module and context have been linked
to the forum instance.