MDL-42609 course: management default missing bits
With the previous commit in this issue, now the course management
UI defaults to show the "default category" courses.

But some bits were missing in order to make that default behavior
fully equivalent with the manual navigation to that category.

This patch:
  - Adds the missing categoryid url param, used by some of the options
    in the UI, like sorting, that was completely broken.
  - Sets the context to the default category, previously was being set to
    system context, incorrectly if now we are defaulting to a category.

That way everything is 100% the same than when we land to the category manually.