MDL-43089 quiz editing: fix RTL
As part of this, I tweaked the LTR style a bit, in a way that is both an
improvement (I hope) and also to reduce the amount of RTL-specific code
required. (Thanks to Fred for spotting some of the problems.)

Also, I changed one course style in boostrapbase, where the RTL rule was
different from the LTR rule.

Finally, I saw some repeated RTL rules in the quiz styles.css, and so
eliminated the the duplicate rules.